Minghui development (#50)
* test pull (#48)

* remove test file

* modify the email module to use the rdas system to send email, move the initiation firebase admin to the star_dev, replaced ses_firebase.py codes by test_ses_firebase.py

* remove test emails and remove '/' from file_paths in the sysvars.py

---------
